% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeomData.R
\name{addGeomData}
\alias{addGeomData}
\title{Put geometry data in a NetCDF-CF File}
\usage{
addGeomData(nc_file, geomData, instance_dim_name, variables = c())
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created. It must already have
an instance dimension.}

\item{geomData}{An object of class \code{SpatialLines}, \code{SpatialPolygons}
or their sf with WGS84 lon in the x coordinate and lat in the y coordinate.
Note that three dimensional geometries are not supported.}

\item{instance_dim_name}{A string to name the instance dimension.  Defaults to "instance"}

\item{variables}{A character vector of variable names that the geometry data
container variable name will be added to.}
}
\description{
Creates a file with point, line or polygon instance data ready for the extended NetCDF-CF timeSeries featuretype format.
Will also add attributes if a sp dataframe object is passed in.
}
\references{
https://github.com/twhiteaker/netCDF-CF-simple-geometry
}
