% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdp_count_time_steps.R
\name{gdp_count_time_steps}
\alias{gdp_count_time_steps}
\title{Get Geo Data Portal time step counter.}
\usage{
gdp_count_time_steps(GDPFile, originDate, expected_times)
}
\arguments{
\item{originDate}{The origin date desired for the time steps to be returned. This is passed to julian as the origin input.}

\item{expected_times}{A rough estimate of how many time steps to expect in the file. It is doubled to preallocate an array so no need to be precise, just order of magnitude.}

\item{gdp_file}{A Geo Data Portal output file containing one statistic, currently no support for multiple statistic output files.}
}
\value{
times A vector of time steps as returned by the Julian function suitable to be passed into a NetCDF file.
}
\description{
A utility function to help figure out how big a GDP file is and get the time steps in it.
}

