% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdp_ids_and_units.R
\name{gdp_ids_and_units}
\alias{gdp_ids_and_units}
\title{Get Geo Data Portal station ids and units}
\usage{
gdp_ids_and_units(GDPFile)
}
\arguments{
\item{gdp_file}{A Geo Data Portal output file containing one statistic, currently no support for multiple statistic output files.}
}
\value{
A named list as follows:\cr
nc_stations contains unique station ids,\cr
nc_data_var_unique_name contains the variables in the file,\cr
unique_units contains the units associated with each variable.
}
\description{
Parses the first three rows of a Geo Data Portal output file and returns a named list ready for use in creation of netCDF DSG files.
}

