% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGmFromPrj.R
\name{getGmFromPrj}
\alias{getGmFromPrj}
\title{Get NetCDF-CF Grid Mapping from Projection}
\usage{
getGmFromPrj(prj)
}
\arguments{
\item{prj}{A proj.4 string as returned from the sp CRS function.}
}
\value{
A named list containing attributes required for that grid_mapping.
}
\description{
This function takes a proj4 string and returns a NetCDF-CF projection container as
a named list of attributes.
}
\details{
https://en.wikibooks.org/wiki/PROJ.4
https://trac.osgeo.org/gdal/wiki/NetCDF_ProjectionTestingStatus
http://cfconventions.org/cf-conventions/cf-conventions.html#appendix-grid-mappings
}
\examples{
prj <- "+proj=longlat +a=6378137 +f=0.00335281066474748 +pm=0 +units=m +no_defs"
grid_mapping <- getGmFromPrj(prj)

}
