% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_instance_data.R
\name{read_instance_data}
\alias{read_instance_data}
\title{Get Dataframe from NetCDF-DSG File}
\usage{
read_instance_data(nc, instance_dim)
}
\arguments{
\item{nc}{A open ncdf4 object.}

\item{instance_dim}{The NetCDF instance/station dimension.}
}
\description{
Gets instance data from a NetCDF-DSG file and returns it in a \code{data.frame}.
This function is intended as a convenience to be used within workflows where
the netCDF file is already open and well understood.
}
