% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_timeseries_dsg.R
\name{read_timeseries_dsg}
\alias{read_timeseries_dsg}
\title{Read timeseries NCDF file}
\usage{
read_timeseries_dsg(nc_file)
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created.}
}
\description{
This reads a timeseries discrete sampling geometry NCDF file
}
\details{
TBD
}
\references{
http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/reference/FeatureDatasets/CFpointImplement.html
}

