% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_stars.R
\name{st_as_stars}
\alias{st_as_stars}
\alias{st_as_stars.ncdfgeom}
\title{Convert ncdfgeom object into stars object.}
\usage{
\method{st_as_stars}{ncdfgeom}(.x, ..., sf_geometry = NA)
}
\arguments{
\item{.x}{Object of class ncdfgeom as returned by read_timeseries_dsg.}

\item{...}{not used.}

\item{sf_geometry}{sf data.frame with geometry and attributes to be added to stars object. 
Must have same number of rows as timeseries instances.}
}
\description{
Convert ncdfgeom object into stars object.
}
