% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_ragged_timeseries_dfg.R
\name{write_ragged_timeseries_dfg}
\alias{write_ragged_timeseries_dfg}
\title{Create ragged array timeseries NCDF file}
\usage{
write_ragged_timeseries_dfg(nc_file, station_names, lats, lons, times, data,
  data_unit = "", data_prec = "double", attributes = list())
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created.}

\item{lats}{Vector of latitudes}

\item{lons}{Vector of longitudes}

\item{times}{A vector of times. Must be of type \code{\link{POSIXct}} or an attempt to
convert it will be made using \code{\link{as.POSIXct(times)}}.}

\item{data}{\code{data.frame} with each column corresponding to a observation. Rows correspond to
samples at different times. nrow must be the same length as times.
Column names are used as names in the NCDF file.}

\item{data_unit}{Character vector of observation units. Length must be the same as number
of columns in \code{data} parameter}

\item{data_prec}{Precision of observation data in NCDF file.
Valid options: 'short' 'integer' 'float' 'double' 'char'.}

\item{attributes}{An optional list of attributes that will be added at the global level.
See details for useful attributes.}
}
\description{
This creates a timeseries discrete sampling features NCDF file
}
\details{
title = "title"
abstract = "history"
provider site = "institution"
provider name ="source"
description = "description"
}
\references{
http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/reference/FeatureDatasets/CFpointImplement.html
}

