% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ragged_timeseries_dsg.R
\name{write_ragged_timeseries_dsg}
\alias{write_ragged_timeseries_dsg}
\title{Create ragged array timeseries NCDF file}
\usage{
write_ragged_timeseries_dsg(nc_file, all_data, data_units = list(),
  data_longnames = list(), attributes = list())
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created.}

\item{all_data}{A data.frame object containing all station and observation data. See 
description for further information}

\item{data_units}{Optional list of observation units.  Names must correspond
to names of observation columns in \code{all_data}. Not required. Columns without 
supplied data unit are given a unit of empty string.}

\item{data_longnames}{Optional named list of observation long names. Names must correspond
to names of observation columns in \code{all_data}. Not required. Columns without 
supplied long name are given a long name of empty string.}

\item{attributes}{An optional list of attributes that will be added at the global level. 
See details for useful attributes.}
}
\description{
This creates a ragged array timeseries discrete sampling features NCDF file
}
\details{
It's challenging to create a function that accepts the style of data that could go into
the ragged array timeseries DSG. This is one attempt, though ideas and feedback are welcome.

The all_data data.frame is the heart of the data passed. It \emph{must} have the following
columns. 
\itemize{
\item{station_name}
\item{lat}
\item{lon}
\item{alt}
\item{time}
\item{observation1}
\item{...}
\item{observationN}
}

observation1 and observationN are just placeholders. Observations should be described with 
a descriptive name. There can be any number of observations, but their names must be unique. 

Attributes allow for addition, non-core metadata to be included. Must be a named list. All
items in the list will be added as global attributes to the NC file with the supplied name/value 
pair. 

Here are some common important attributes that could be added.
\itemize{
\item{title = "title"}
\item{abstract = "history"}
\item{provider site = "institution"}
\item{provider name ="source"}
\item{description = "description"}
}
}
\examples{
\dontrun{
library(soilmoisturetools)
ok = ok_data()
ok_meta = ok_sites_metadata(ok$station)[, c('station', 'latitude', 
                                           'longitude', 'elevation')]
all_data = merge(ok, ok_meta)

names(all_data) = c('station_name', 'time', 'soil_moisture', 
                   'soil_moisture_depth', 'lat', 'lon', 'alt')
write_ragged_timeseries_dsg('test.nc', 
                            all_data,list(soil_moisture='\%', 
                            soil_moisture_depth='inches'))

}
}

