% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromNCDFSG.R
\name{FromNCDFSG}
\alias{FromNCDFSG}
\title{Convert NetCDF to sp objects}
\usage{
FromNCDFSG(nc_file)
}
\arguments{
\item{nc_file}{A string file path to the nc file to be read.}
}
\value{
sp object containing spatial geometry of type found in the NetCDF-CF DSG file.
}
\description{
Attemps to convert a NetCDF-CF DSG Simple Geometry file into an sp object.
}
\references{
https://github.com/twhiteaker/netCDF-CF-simple-geometry
}
