% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gridmapping.R
\name{get_gridmapping}
\alias{get_gridmapping}
\title{Get NetCDF-CF grid mapping from projection}
\usage{
get_gridmapping(prj)
}
\arguments{
\item{prj}{A proj.4 string as returned from rgdal::CRSargs() or
sf::st_crs() or in the @proj4string slot of a sp object.}
}
\value{
A named list containing attributes required for that grid_mapping.
}
\description{
Takes a proj4 string and returns a NetCDF-CF projection as
a named list of attributes.
}
\examples{
sample_data <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))
get_gridmapping(sf::st_crs(sample_data)$proj4string)
get_gridmapping(sf::st_crs("+init=epsg:5070")$proj4string)

sample_data <- sf::as_Spatial(sample_data)
get_gridmapping(sample_data@proj4string)
get_gridmapping("+proj=longlat +a=6378137 +f=0.00335281066474748 +pm=0 +no_defs")

}
\references{
\enumerate{
  \item \url{https://en.wikibooks.org/wiki/PROJ.4}
  \item \url{https://trac.osgeo.org/gdal/wiki/NetCDF_ProjectionTestingStatus}
  \item \url{http://cfconventions.org/cf-conventions/cf-conventions.html#appendix-grid-mappings}
}
}
