% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geometry.R
\name{write_geometry}
\alias{write_geometry}
\title{Write sp or sf geometries and attributes to NetCDF-CF}
\usage{
write_geometry(nc_file, geomData = NULL, instance_dim_name = NULL,
  lats = NULL, lons = NULL, variables = list())
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created.}

\item{geomData}{An object of class \code{SpatialLines}, \code{SpatialPolygons}
or their sf equivalents with WGS84 lon in the x coordinate and lat in the y coordinate.
Note that three dimensional geometries are not supported.}

\item{instance_dim_name}{Not required if adding geometry to a NetCDF-CF Discrete Sampling Geometries 
timeSeries file. For a new file, will use package default -- "instance" -- if not supplied.}

\item{lats}{Vector of WGS84 latitudes}

\item{lons}{Vector of WGS84 longitudes}

\item{variables}{If a an existing netcdf files is provided, this list of strings is used
to add the geometry container attribute to the named existing variables.}
}
\description{
Creates a file with point, line or polygon instance data ready for the extended NetCDF-CF timeSeries featuretype format.
Will also add attributes if a sp dataframe object is passed in.
}
\references{
https://github.com/twhiteaker/netCDF-CF-simple-geometry
}
