% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_point_dsg.R
\name{write_point_dsg}
\alias{write_point_dsg}
\title{Write NetCDF-CF point featuretype}
\usage{
write_point_dsg(nc_file, lats, lons, alts, times, data,
  data_units = rep("", ncol(data)), feature_names = NULL, ...)
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created.}

\item{lats}{Vector of latitudes}

\item{lons}{Vector of longitudes}

\item{alts}{Vector of altitudes}

\item{times}{Vector of times, one per point, if length is one, will use the same time for each point. 
Must be of type \code{POSIXct} or an attempt to convert it will be made using \code{as.POSIXct(times)}.}

\item{data}{\code{data.frame} with each column corresponding to a observation. Column 
names are used as names in the NCDF file}

\item{data_units}{Character vector of observation units. Length must be the same as number 
of columns in \code{data} parameter}

\item{feature_names}{\code{vector} of identifiers for features or stations.}

\item{...}{additional arguments to be passed on nc_create.}
}
\description{
This creates a simple point data discrete sampling features NCDF file. Returns the created filename. 
Can pass in netcdf creation options like force_v4 to pass on.
}
\references{
http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/reference/FeatureDatasets/CFpointImplement.html
}
